# docs/swagger.yaml
basePath: /
definitions:
  github_com_mrexmelle_connect-emp_internal_dto.ServiceError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  internal_grading.DeleteResponseDto:
    properties:
      error:
        $ref: '#/definitions/github_com_mrexmelle_connect-emp_internal_dto.ServiceError'
    type: object
  internal_grading.GetResponseDto:
    properties:
      data:
        $ref: '#/definitions/internal_grading.ViewEntity'
      error:
        $ref: '#/definitions/github_com_mrexmelle_connect-emp_internal_dto.ServiceError'
    type: object
  internal_grading.PatchRequestDto:
    properties:
      fields:
        additionalProperties: true
        type: object
    type: object
  internal_grading.PatchResponseDto:
    properties:
      error:
        $ref: '#/definitions/github_com_mrexmelle_connect-emp_internal_dto.ServiceError'
    type: object
  internal_grading.PostRequestDto:
    properties:
      ehid:
        type: string
      end_date:
        type: string
      grade:
        type: string
      start_date:
        type: string
    type: object
  internal_grading.PostResponseDto:
    properties:
      data:
        $ref: '#/definitions/internal_grading.ViewEntity'
      error:
        $ref: '#/definitions/github_com_mrexmelle_connect-emp_internal_dto.ServiceError'
    type: object
  internal_grading.ViewEntity:
    properties:
      ehid:
        type: string
      end_date:
        type: string
      grade:
        type: string
      id:
        type: integer
      start_date:
        type: string
    type: object
  internal_profile.GetResponseDto:
    properties:
      data:
        $ref: '#/definitions/libauthxc.ProfileEntity'
      error:
        $ref: '#/definitions/github_com_mrexmelle_connect-emp_internal_dto.ServiceError'
    type: object
  internal_titling.DeleteResponseDto:
    properties:
      error:
        $ref: '#/definitions/github_com_mrexmelle_connect-emp_internal_dto.ServiceError'
    type: object
  internal_titling.GetResponseDto:
    properties:
      data:
        $ref: '#/definitions/internal_titling.ViewEntity'
      error:
        $ref: '#/definitions/github_com_mrexmelle_connect-emp_internal_dto.ServiceError'
    type: object
  internal_titling.PatchRequestDto:
    properties:
      fields:
        additionalProperties: true
        type: object
    type: object
  internal_titling.PatchResponseDto:
    properties:
      error:
        $ref: '#/definitions/github_com_mrexmelle_connect-emp_internal_dto.ServiceError'
    type: object
  internal_titling.PostRequestDto:
    properties:
      ehid:
        type: string
      end_date:
        type: string
      start_date:
        type: string
      title:
        type: string
    type: object
  internal_titling.PostResponseDto:
    properties:
      data:
        $ref: '#/definitions/internal_titling.ViewEntity'
      error:
        $ref: '#/definitions/github_com_mrexmelle_connect-emp_internal_dto.ServiceError'
    type: object
  internal_titling.ViewEntity:
    properties:
      ehid:
        type: string
      end_date:
        type: string
      id:
        type: integer
      start_date:
        type: string
      title:
        type: string
    type: object
  libauthxc.ProfileEntity:
    properties:
      dob:
        type: string
      ehid:
        type: string
      email_address:
        type: string
      employee_id:
        type: string
      name:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8082
info:
  contact:
    email: mrexmelle@gmail.com
  description: Employee API for Connect.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Connect Employee API
  version: 0.1.0
paths:
  /gradings:
    post:
      consumes:
      - application/json
      description: Post a new gradings
      parameters:
      - description: Grading Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_grading.PostRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success Response
          schema:
            $ref: '#/definitions/internal_grading.PostResponseDto'
        "400":
          description: BadRequest
        "500":
          description: InternalServerError
      tags:
      - Gradings
  /gradings/{id}:
    delete:
      description: Delete a grading
      parameters:
      - description: Grading ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Response
          schema:
            $ref: '#/definitions/internal_grading.DeleteResponseDto'
        "400":
          description: BadRequest
        "500":
          description: InternalServerError
      tags:
      - Gradings
    get:
      description: Get a grading
      parameters:
      - description: Grading ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Response
          schema:
            $ref: '#/definitions/internal_grading.GetResponseDto'
        "400":
          description: BadRequest
        "500":
          description: InternalServerError
      tags:
      - Gradings
    patch:
      consumes:
      - application/json
      description: Patch a grading
      parameters:
      - description: Grading ID
        in: path
        name: id
        required: true
        type: string
      - description: Grading Patch Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_grading.PatchRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success Response
          schema:
            $ref: '#/definitions/internal_grading.PatchResponseDto'
        "400":
          description: BadRequest
        "500":
          description: InternalServerError
      tags:
      - Gradings
  /profiles/{ehid}:
    get:
      description: Get a profile
      parameters:
      - description: EHID
        in: path
        name: ehid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Response
          schema:
            $ref: '#/definitions/internal_profile.GetResponseDto'
        "400":
          description: BadRequest
        "500":
          description: InternalServerError
      tags:
      - Profiles
  /titlings:
    post:
      consumes:
      - application/json
      description: Post a new titlings
      parameters:
      - description: Titling Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_titling.PostRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success Response
          schema:
            $ref: '#/definitions/internal_titling.PostResponseDto'
        "400":
          description: BadRequest
        "500":
          description: InternalServerError
      tags:
      - Titlings
  /titlings/{id}:
    delete:
      description: Delete a titling
      parameters:
      - description: Titling ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Response
          schema:
            $ref: '#/definitions/internal_titling.DeleteResponseDto'
        "400":
          description: BadRequest
        "500":
          description: InternalServerError
      tags:
      - Titlings
    get:
      description: Get a titling
      parameters:
      - description: Titling ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Response
          schema:
            $ref: '#/definitions/internal_titling.GetResponseDto'
        "400":
          description: BadRequest
        "500":
          description: InternalServerError
      tags:
      - Titlings
    patch:
      consumes:
      - application/json
      description: Patch a titling
      parameters:
      - description: Titling ID
        in: path
        name: id
        required: true
        type: string
      - description: Titling Patch Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_titling.PatchRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success Response
          schema:
            $ref: '#/definitions/internal_titling.PatchResponseDto'
        "400":
          description: BadRequest
        "500":
          description: InternalServerError
      tags:
      - Titlings
swagger: "2.0"